improve WT_ITEM internal interface; move InnerCursor into module root
